--- .github/workflows/1_tests.yml
@@ -1,45 +1,0 @@
-name: Tests
-
-on:
-  push:
-    branches: [ main ]
-  pull_request:
-    branches: [ main ]
-  workflow_dispatch:
-
-jobs:
-  test:
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        os: [ubuntu-latest]
-        python-version: ['3.7', '3.8', '3.9', '3.10']
-        include:
-          - os: windows-latest
-            python-version: '3.7'
-          - os: macos-latest
-            python-version: '3.7'
-    timeout-minutes: 10
-
-    steps:
-    - name: Check out code repository
-      uses: actions/checkout@v2
-
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v2
-      with:
-        python-version: ${{ matrix.python-version }}
-
-    - name: Install dependencies and package
-      run: |
-        python -m ensurepip
-        python -m pip install --upgrade pip
-        python -m pip install poetry==${{ matrix.poetry-version }}
-
-    - name: Install dependencies
-      shell: bash
-      run: python -m poetry install
-
-    - name: Test with pytest
-      run: |
-        python -m poetry run python -m pytest
