--- .github/workflows/2_coverage.yml
@@ -1,46 +1,0 @@
-name: Code coverage
-
-on:
-  push:
-    branches: [ main ]
-  pull_request:
-    branches: [ main ]
-  workflow_dispatch:
-
-jobs:
-  run:
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        os: [ubuntu-latest]
-        python-version: ['3.7']
-    env:
-      OS: ${{ matrix.os }}
-      PYTHON: ${{ matrix.python-version }}
-    timeout-minutes: 10
-
-    steps:
-    - name: Check out code repository
-      uses: actions/checkout@v2
-
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v2
-      with:
-        python-version: ${{ matrix.python-version }}
-
-    - name: Install poetry ${{ matrix.poetry-version }}
-      run: |
-        python -m ensurepip
-        python -m pip install --upgrade pip
-        python -m pip install poetry==${{ matrix.poetry-version }}
-
-    - name: Install dependencies
-      shell: bash
-      run: python -m poetry install
-
-    - name: Generate coverage report
-      run: |
-        python -m poetry run python -m pytest --cache-clear --cov=app tests/ > pytest-coverage.txt
-
-    - name: Comment coverage on PR
-      uses: coroo/pytest-coverage-commentator@v1.0.2
