--- .github/workflows/3_linting.yml
@@ -1,51 +1,0 @@
-name: Lint and format code
-
-on:
-  push:
-    branches: [ main ]
-  pull_request:
-    branches: [ main ]
-  workflow_dispatch:
-
-jobs:
-  lint:
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        os: [ubuntu-latest]
-        python-version: ['3.7']
-    env:
-      OS: ${{ matrix.os }}
-      PYTHON: ${{ matrix.python-version }}
-    timeout-minutes: 10
-
-    steps:
-    - name: Check out code repository
-      uses: actions/checkout@v2
-
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v2
-      with:
-        python-version: ${{ matrix.python-version }}
-
-    - name: Install poetry ${{ matrix.poetry-version }}
-      run: |
-        python -m ensurepip
-        python -m pip install --upgrade pip
-        python -m pip install poetry==${{ matrix.poetry-version }}
-
-    - name: Install dependencies
-      shell: bash
-      run: python -m poetry install
-
-    - name: Lint with pylint
-      run: |
-        pylint actionscicd
-        pylint tests
-
-    - name: Run Black check
-      uses: psf/black@stable
-
-    - name: Check type hints with mypy
-      run: |
-        poetry run mypy --strict dates
