--- .github/workflows/6_pypi_release.yml
@@ -1,37 +1,0 @@
-name: Upload release to PyPI
-
-on:
-  release:
-    types: [created]
-
-jobs:
-  deploy:
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        os: [ubuntu-latest]
-        python-version: [3.8]
-        poetry-version: [1.1.4]
-    env:
-      OS: ${{ matrix.os }}
-      PYTHON: ${{ matrix.python-version }}
-    timeout-minutes: 10
-
-    steps:
-    - name: Check out code repository
-      uses: actions/checkout@v2
-
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v2
-      with:
-        python-version: ${{ matrix.python-version }}
-
-    - name: Build and publish to pypi
-      uses: JRubics/poetry-publish@v1.10
-      with:
-        pypi_token: ${{ secrets.PYPI_TEST_TOKEN }}
-        repository_name: "testpypi"
-        repository_url: "https://test.pypi.org/legacy/"
-        ignore_dev_requirements: "yes"
-        python_version: "${{ matrix.python-version }}"
-        poetry_version: "==${{ matrix.poetry-version }}"
