+++ Dockerfile
@@ -2,16 +2,17 @@ ARG APP_DIR="/tmp"
 
 FROM python:3.8
 
-ARG FUNCTION_DIR
+ARG APP_DIR
+WORKDIR ${APP_DIR}
 
 # Install python packages
 RUN pip install 'poetry==1.1.4'
-COPY poetry.lock pyproject.toml ${FUNCTION_DIR}/
+COPY poetry.lock pyproject.toml ${APP_DIR}/
 RUN poetry config virtualenvs.create false && poetry install --no-interaction --no-ansi
 
 # Copy function code
 COPY . .
 
-ENV PYTHONPATH "${PYTHONPATH}:${FUNCTION_DIR}"
+ENV PYTHONPATH "${PYTHONPATH}:${APP_DIR}"
 
-CMD [ "poetry", "python", "main.py" ]
+ENTRYPOINT [ "poetry", "run", "python", "main.py" ]
