+++ README.md
@@ -1,1 +1,87 @@
-# hai-dates-tech-task
+# Harrison AI Dates Tech Task
+> This module computes the number of days between two dates.
+
+## Table of Contents
+* [General Info](#general-information)
+* [Technologies Used](#technologies-used)
+* [Features](#features)
+* [Screenshots](#screenshots)
+* [Setup](#setup)
+* [Usage](#usage)
+* [Project Status](#project-status)
+* [Room for Improvement](#room-for-improvement)
+* [Acknowledgements](#acknowledgements)
+* [Contact](#contact)
+<!-- * [License](#license) -->
+
+
+## General Information
+The aim of the project is to compute the difference between two dates from scratch. I.e. without importing or otherwise building on existing packages for date processing.
+Timezones were not considered and the dates were expected to be of format `YYYY-MM-DD` and based of the [Gregorian calendar](https://en.wikipedia.org/wiki/Gregorian_calendar).
+
+
+## Technologies Used
+- [Poetry](https://python-poetry.org/) - Python package and dependency management
+- [Docker](https://www.docker.com/) - Containerisation tool
+- [VSCode Remote Containers](https://code.visualstudio.com/docs/remote/containers) - Containerised development
+- [GitHub Actions](https://github.com/features/actions) - CI/CD tool
+
+
+## Features
+- Python module to calculate date difference (`dates/date_diff.py`) with tests (`dates/tests/`)
+- Dockerfile for reproducibility across dev and 'prod' environments
+- Awesome feature 3
+
+
+## Screenshots
+![Example screenshot](./img/screenshot.png)
+<!-- If you have screenshots you'd like to share, include them here. -->
+
+
+## Setup
+What are the project requirements/dependencies? Where are they listed? A requirements.txt or a Pipfile.lock file perhaps? Where is it located?
+
+Proceed to describe how to install / setup one's local environment / get started with the project.
+
+
+## Usage
+How does one go about using it?
+Provide various use cases and code examples here.
+
+`write-your-code-here`
+
+
+## Project Status
+Project is: _in progress_ / _complete_ / _no longer being worked on_. If you are no longer working on it, provide reasons why.
+
+
+## Room for Improvement
+Include areas you believe need improvement / could be improved. Also add TODOs for future development.
+
+Room for improvement:
+- Improvement to be done 1
+- Improvement to be done 2
+
+To do:
+- Feature to be added 1
+- Feature to be added 2
+
+- Documentation (Sphynx)
+
+
+## Acknowledgements
+Give credit here.
+- This project was inspired by...
+- This project was based on [this tutorial](https://www.example.com).
+- Many thanks to...
+
+
+## Contact
+Created by [@flynerdpl](https://www.flynerd.pl/) - feel free to contact me!
+
+
+<!-- Optional -->
+<!-- ## License -->
+<!-- This project is open source and available under the [... License](). -->
+
+<!-- You don't have to include all sections - just the one's relevant to your project -->
