+++ build.sh
@@ -1,0 +1,1 @@
+docker build -t hai-dates .
