+++ main.py
@@ -1,6 +1,11 @@
+"""Main entrypoint"""
 import fire
+from loguru import logger
 
-from dates.dates import date_diff
+from dates.date_diff import date_diff
 
 if __name__ == "__main__":
-    fire.Fire(date_diff)
+    try:
+        fire.Fire(date_diff)
+    except fire.core.FireExit as e:
+        logger.error(e.trace)
