+++ pyproject.toml
@@ -7,6 +7,7 @@ authors = ["Elmi de Lange"]
 [tool.poetry.dependencies]
 python = "^3.8"
 fire = "^0.4.0"
+loguru = "^0.6.0"
 
 [tool.poetry.dev-dependencies]
 black = "^22.1.0"
